+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [23-08-13]
+
+## Changed
+
+- Layout der PDF Doku
+- Site has now City ZIP Street
+- Daschboard shows Sites
+
+
+
 ## [23-08-12]
 
 ## Added
@@ -34,7 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-## [23-08-05]
+## [23-08-06]
 
 ## Added